File restructure
Created directories for organizing the project. moved files around
